requirements.txt
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.37.1
langchain==0.2.14
langchain-community==0.2.12
